.json now saves code blocks, and moved more stuff to editorJsonHandler.js
