+++ src/game/commonstrings.js
@@ -72,5 +72,5 @@ export class TaskTypes {
     static instruction = "instructions"; //read instructions
     static multipleChoice = "multiple-choice"; //Choose correct answer
     static codeFixing = "code-fixing"; //Correct the code (not implemented)
-    static codeBlockMoving = "code-block-Moving"; //Move code blocks to finish the level
+    static codeBlockMoving = "code-block-moving"; //Move code blocks to finish the level
 }
